requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
numpy>=1.22
torch>=1.13
Pillow>=9.0
openai>=1.3
sentence-transformers>=2.2